apps/api/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-dotenv>=1.0.0
duckdb>=0.9.0
pandas>=2.0.0
pyarrow>=14.0.0
numpy>=1.24.0
httpx>=0.24.0
pytest>=7.0.0
pytest-cov>=4.1.0
black>=23.3.0
ruff>=0.0.1 